use Model and View
