+++ js/narrativeModel.js
@@ -1,0 +1,12 @@
+define([
+    'core/js/adapt',
+    'core/js/models/itemsModel'
+], function(Adapt, ItemsModel) {
+
+    var NarrativeModel = ItemsModel.extend({
+        
+    });
+
+    return NarrativeModel;
+
+});
